cleaning up the goemetry class
a lot of things were cleaned up in the geometry class. The class now includes a generateTextureCoordinates function that does just that. It is no longer handled in the geometry extention classes. The Tessellator.Polygon class is now fixed, and creating geometry with the polygon primitive will work properly. One more side note: the constant class has been removed and all constants are now interegs like they are in the webGL counterpart. A lot of cleaning up this week, The gods are smileing on this one.